Fixed renderer, ready to move on to more things

+++ game-engine/game-engine/src/services/Services.java
@@ -50,7 +50,7 @@ public class Services {
 	}
 	
 	public static void cleanup(){
-		ModelData data=new ModelData(new float[0], new short[0]);
+		ModelData data=new ModelData(new float[0], new int[0]);
 		data.cleanup();
 		
 		getShader().cleanup();
